# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BookBuddy API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NotFoundError {
  message: string;
}

export interface User {
  id: number;
  displayName: string;
  dailyGoalMinutes: number;
  reminderEnabled: boolean;
  reminderTime?: string | null;
  burnoutWindowStart?: string | null;
  burnoutWindowEnd?: string | null;
  darkMode: boolean;
  createdAt: string;
}

export interface OnboardUserBody {
  displayName: string;
  bookTitle: string;
  bookAuthor?: string | null;
  bookTotalPages: number;
  currentPage?: number;
  dailyGoalMinutes?: number;
  reminderEnabled?: boolean;
  reminderTime?: string | null;
  burnoutWindowStart?: string | null;
  burnoutWindowEnd?: string | null;
  darkMode?: boolean;
}

export interface UpdateUserBody {
  displayName?: string;
  dailyGoalMinutes?: number;
  reminderEnabled?: boolean;
  reminderTime?: string | null;
  burnoutWindowStart?: string | null;
  burnoutWindowEnd?: string | null;
  darkMode?: boolean;
}

export interface Book {
  id: number;
  title: string;
  author?: string | null;
  totalPages: number;
  currentPage: number;
  percentComplete: number;
  isCompleted: boolean;
  completedAt?: string | null;
  lastReadAt?: string | null;
  totalSessions: number;
  isCurrent: boolean;
  createdAt: string;
}

export interface CreateBookBody {
  title: string;
  author?: string | null;
  totalPages: number;
  currentPage?: number;
  setAsCurrent?: boolean;
}

export interface UpdateBookBody {
  title?: string;
  author?: string | null;
  totalPages?: number;
  currentPage?: number;
  isCurrent?: boolean;
}

export interface ReadingSession {
  id: number;
  bookId: number;
  bookTitle: string;
  pagesRead: number;
  minutesRead: number;
  startPage: number;
  endPage: number;
  sessionDate: string;
  notes?: string | null;
  createdAt: string;
}

export interface CreateSessionBody {
  bookId: number;
  pagesRead: number;
  minutesRead: number;
  startPage: number;
  endPage: number;
  notes?: string | null;
}

export interface DashboardSummary {
  currentStreak: number;
  longestStreak: number;
  weeklyActiveDays: number;
  weeklyGoalDays: number;
  momentumLabel: string;
  totalPagesRead: number;
  totalMinutesRead: number;
  lastReadAt?: string | null;
  daysSinceLastRead?: number | null;
  isRecovering: boolean;
  currentBook?: Book | null;
  todayMinutesRead: number;
  todayPagesRead: number;
  todayGoalMet: boolean;
  encouragementMessage: string;
}

export interface DayActivity {
  date: string;
  dayLabel: string;
  isActive: boolean;
  isToday: boolean;
  isFuture: boolean;
  pagesRead: number;
  minutesRead: number;
}

export interface WeeklySummary {
  week: DayActivity[];
  activeDays: number;
  totalPagesThisWeek: number;
  totalMinutesThisWeek: number;
}

export type ReadingHistorySessionsByDate = { [key: string]: ReadingSession[] };

export interface MonthlyStats {
  totalSessions: number;
  totalPages: number;
  totalMinutes: number;
  activeDays: number;
}

export interface ReadingHistory {
  month: string;
  activeDates: string[];
  sessionsByDate: ReadingHistorySessionsByDate;
  monthlyStats: MonthlyStats;
}

export interface Settings {
  userId: number;
  dailyGoalMinutes: number;
  reminderEnabled: boolean;
  reminderTime?: string | null;
  burnoutWindowStart?: string | null;
  burnoutWindowEnd?: string | null;
  darkMode: boolean;
}

export interface UpdateSettingsBody {
  dailyGoalMinutes?: number;
  reminderEnabled?: boolean;
  reminderTime?: string | null;
  burnoutWindowStart?: string | null;
  burnoutWindowEnd?: string | null;
  darkMode?: boolean;
}

export type ListSessionsParams = {
  bookId?: number;
  limit?: number;
  offset?: number;
};

export type GetRecentSessionsParams = {
  limit?: number;
};

export type GetReadingHistoryParams = {
  /**
   * YYYY-MM format
   */
  month?: string;
};