# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BookBuddy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List all books
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"

    post:
      operationId: createBook
      tags: [books]
      summary: Add a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookBody"
      responses:
        "201":
          description: Created book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a specific book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

    put:
      operationId: updateBook
      tags: [books]
      summary: Update a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookBody"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/{id}/complete:
    post:
      operationId: completeBook
      tags: [books]
      summary: Mark book as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/current:
    get:
      operationId: getCurrentBook
      tags: [books]
      summary: Get the currently active book
      responses:
        "200":
          description: Current book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: No current book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
...
  /dashboard/history:
    get:
      operationId: getReadingHistory
      tags: [dashboard]
      summary: Get reading history by date with sessions
      parameters:
        - name: month
          in: query
          schema:
            type: string
            description: "YYYY-MM format"
      responses:
        "200":
          description: Reading history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingHistory"
...
components:
  schemas:
...
    NotFoundError:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
          nullable: true
        totalPages:
          type: integer
        currentPage:
          type: integer
        percentComplete:
          type: number
        isCompleted:
          type: boolean
        completedAt:
          type: string
          format: date-time
          nullable: true
        lastReadAt:
          type: string
          format: date-time
          nullable: true
        totalSessions:
          type: integer
        isCurrent:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - totalPages
        - currentPage
        - percentComplete
        - isCompleted
        - totalSessions
        - isCurrent
        - createdAt
    CreateBookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
          nullable: true
        totalPages:
          type: integer
        currentPage:
          type: integer
          default: 0
        setAsCurrent:
          type: boolean
          default: true
      required:
        - title
        - totalPages
    UpdateBookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
          nullable: true
        totalPages:
          type: integer
        currentPage:
          type: integer
        isCurrent:
          type: boolean
    ReadingSession:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        bookTitle:
          type: string
        pagesRead:
          type: integer
        minutesRead:
          type: integer
        startPage:
          type: integer
        endPage:
          type: integer
        sessionDate:
          type: string
          format: date
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - bookId
        - bookTitle
        - pagesRead
        - minutesRead
        - startPage
        - endPage
        - sessionDate
        - createdAt
...
    ReadingHistory:
      type: object
      properties:
        month:
          type: string
        activeDates:
          type: array
          items:
            type: string
            format: date
        sessionsByDate:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/ReadingSession"
        monthlyStats:
          $ref: "#/components/schemas/MonthlyStats"
      required:
        - month
        - activeDates
        - sessionsByDate
        - monthlyStats
    MonthlyStats:
      type: object
      properties:
        totalSessions:
          type: integer
        totalPages:
          type: integer
        totalMinutes:
          type: integer
        activeDays:
          type: integer
      required:
        - totalSessions
        - totalPages
        - totalMinutes
        - activeDays
...